pyproject.toml
[project]
name = "clawstrike"
version = "0.1.0"
description = "Security guardrails layer for OpenClaw AI agent"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=3.0.1",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "structlog>=24.0",
    "torch>=2.10.0",
    "transformers>=5.2.0",
    "typer>=0.12",
]

[project.scripts]
clawstrike = "clawstrike.cli:app"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/clawstrike"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
src = ["src"]
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.2",
]
